Possible to run without front proxy

+++ 8.0/build_and_run.sh
@@ -15,15 +15,22 @@ IMAGE=ubuntudoo-it:8
 docker stop $CONTAINER
 docker rm $CONTAINER
 docker build -t $IMAGE .
-docker run -d -v $(pwd):/etc/odoo --name odoo8 --link db:db -t ubuntudoo-it:8
+
+if [ "$1" = "no_proxy" ]; then
+    docker run -d -v $(pwd):/etc/odoo -p 80:8069 --name odoo8 --link db:db -t ubuntudoo-it:8
+else
+    docker run -d -v $(pwd):/etc/odoo --name odoo8 --link db:db -t ubuntudoo-it:8
+fi
+
 
 # front proxy
-
+if [ "$1" = "no_proxy" ]; then
 docker stop front_proxy
 docker build front_proxy -f front_proxy/Dockerfile -t front_proxy;
 docker run -d -v /etc/letsencrypt:/etc/letsencrypt -p 80:80 -p 443:443 \
         --name front_proxy --env-file ./front_proxy/env.list \
         --link odoo8:odoo8 --link db:db --privileged --rm front_proxy
+fi
 
 IP=$(echo $(hostname -I) | cut -d' ' -f 1)
 echo "App is running on:"
